fix(pkinit): allow full uint32 PKAuthenticator nonce range
PKINIT nonces can use the full uint32 range, but
the current code stores them in an int and
generates them with `Intn()`, which cannot
represent `MaxUint32` on all plats and also
excludes that value.

Store the nonce as int64 and generate it from
`Uint32()` so ASN.1 encoding can round-trip values
through the full uint32 range.

Fixes #8

+++ pkinit/asn1.go
@@ -97,7 +97,7 @@ type PKAuthenticator struct {
 	// asn1
 	CUSec    int       `asn1:"tag:0,explicit"`
 	CTime    time.Time `asn1:"tag:1,explicit,generalized"`
-	Nonce    int       `asn1:"tag:2,explicit"`
+	Nonce    int64     `asn1:"tag:2,explicit"`
 	Checksum []byte    `asn1:"tag:3,explicit,optional"`
 }
